+++ src/traits/inflect.rs
@@ -7,14 +7,18 @@ use crate::{
 	NiceU16,
 	NiceU32,
 	NiceU64,
+	NiceWrapper,
 };
-use std::num::{
-	NonZeroU8,
-	NonZeroU16,
-	NonZeroU32,
-	NonZeroU64,
-	NonZeroUsize,
-	NonZeroU128,
+use std::{
+	fmt,
+	num::{
+		NonZeroU8,
+		NonZeroU16,
+		NonZeroU32,
+		NonZeroU64,
+		NonZeroUsize,
+		NonZeroU128,
+	},
 };
 
 
@@ -48,7 +52,7 @@ pub trait Inflection: Sized + Copy + PartialEq {
 /// This extends the `Inflection` trait for types which can be represented as
 /// one of the `NiceU*` types, and their signed equivalents (minus signs will
 /// be prepended as necessary), i.e. `i/u/NonZeroU 8–64`.
-pub trait NiceInflection: Inflection {
+pub trait NiceInflection<const S: usize>: Inflection {
 	/// # Inflect a String (Prefixed w/ Value)
 	///
 	/// This is like [`Inflection::inflect`], but prefixes the output with a
@@ -58,13 +62,71 @@ pub trait NiceInflection: Inflection {
 	///
 	/// ```
 	/// use dactyl::traits::NiceInflection;
+	///
 	/// assert_eq!(
-	///     3283.nice_inflect("book", "books"),
-	///     "3,283 books"
+	///     3283.nice_inflect("book", "books").to_string(),
+	///     "3,283 books",
+	/// );
+	///
+	/// // The return type implements fmt::Display, so you can do things
+	/// // like this too:
+	/// assert_eq!(
+	///     format!("I have {}!", 3283.nice_inflect("book", "books")),
+	///     "I have 3,283 books!",
 	/// );
 	/// ```
-	fn nice_inflect<S>(self, singular: S, plural: S) -> String
-	where S: AsRef<str>;
+	fn nice_inflect<'a>(self, singular: &'a str, plural: &'a str) -> NiceInflected<'a, S>;
+}
+
+
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+/// # Nice Inflection Wrapper.
+///
+/// This struct serves as the return type for [`NiceInflection::nice_inflect`].
+/// It implements [`Display`](fmt::Display) so can be chucked straight into a
+/// formatting pattern or converted to a string via `to_string()`.
+///
+/// ## Examples
+///
+/// ```
+/// use dactyl::traits::NiceInflection;
+///
+/// assert_eq!(
+///     format!(
+///         "I have eaten {} and {}!",
+///          1001.nice_inflect("hotdog", "hotdogs"),
+///          1.nice_inflect("hamburger", "hamburger"),
+///     ),
+///     "I have eaten 1,001 hotdogs and 1 hamburger!",
+/// );
+/// ```
+pub struct NiceInflected<'a, const S: usize> {
+	/// # Negative?
+	neg: bool,
+
+	/// # The Number.
+	nice: NiceWrapper<S>,
+
+	/// # The Inflected Text.
+	unit: &'a str,
+}
+
+impl<'a, const S: usize> fmt::Display for NiceInflected<'a, S> {
+	#[inline]
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		// Add a minus sign to the start if negative.
+		if self.neg { f.write_str("-")?; }
+
+		// Print the number.
+		f.write_str(self.nice.as_str())?;
+
+		// Add a space.
+		f.write_str(" ")?;
+
+		// And print the text value.
+		f.write_str(self.unit)
+	}
 }
 
 
@@ -105,72 +167,66 @@ macro_rules! inflect {
 /// # Helper: Generate `Inflection` and `NiceInflection` impls.
 macro_rules! inflect_nice {
 	// Unsigned.
-	($ty:ty, $nice:ty) => (
-		impl NiceInflection for $ty {
+	($size:literal, $ty:ty, $nice:ty) => (
+		impl NiceInflection<$size> for $ty {
 			#[inline]
 			/// # Inflect a String.
-			fn nice_inflect<S>(self, singular: S, plural: S) -> String
-			where S: AsRef<str> {
-				[
-					<$nice>::from(self).as_str(),
-					" ",
-					self.inflect(singular.as_ref(), plural.as_ref()),
-				].concat()
+			fn nice_inflect<'a>(self, singular: &'a str, plural: &'a str) -> NiceInflected<'a, $size> {
+				let nice = <$nice>::from(self);
+				let unit = self.inflect(singular, plural);
+				NiceInflected { neg: false, nice, unit }
 			}
 		}
 	);
 
 	// Signed.
-	($ty:ty, $nice:ty, $cast:ident) => (
-		impl NiceInflection for $ty {
+	($size:literal, $ty:ty, $nice:ty, $cast:ident) => (
+		impl NiceInflection<$size> for $ty {
 			#[inline]
 			/// # Inflect a String.
-			fn nice_inflect<S>(self, singular: S, plural: S) -> String
-			where S: AsRef<str> {
-				[
-					if self < 0 { "-" } else { "" },
-					<$nice>::from(self.$cast()).as_str(),
-					" ",
-					self.inflect(singular.as_ref(), plural.as_ref()),
-				].concat()
+			fn nice_inflect<'a>(self, singular: &'a str, plural: &'a str) -> NiceInflected<'a, $size> {
+				let neg = self < 0;
+				let nice = <$nice>::from(self.$cast());
+				let unit = self.inflect(singular, plural);
+				NiceInflected { neg, nice, unit }
 			}
 		}
 	);
 
 	// Unsigned, both impls.
-	($ty:ty, $nice:ty, $one:literal) => (
+	($size:literal, $ty:ty, $nice:ty, $one:literal) => (
 		inflect!($ty, $one);
-		inflect_nice!($ty, $nice);
+		inflect_nice!($size, $ty, $nice);
 	);
 
 	// Nonzero, both impls.
-	($ty:ty, $nice:ty, $one:expr) => (
+	($size:literal, $ty:ty, $nice:ty, $one:expr) => (
 		inflect!($ty, $one);
-		inflect_nice!($ty, $nice);
+		inflect_nice!($size, $ty, $nice);
 	);
 
 	// Signed, both impls.
-	($ty:ty, $nice:ty, $one:literal, $cast:ident) => (
+	($size:literal, $ty:ty, $nice:ty, $one:literal, $cast:ident) => (
 		inflect!($ty, $one, $cast);
-		inflect_nice!($ty, $nice, $cast);
+		inflect_nice!($size, $ty, $nice, $cast);
 	);
 }
 
-inflect_nice!(u8, NiceU8, 1);
-inflect_nice!(u16, NiceU16, 1);
-inflect_nice!(u32, NiceU32, 1);
-inflect_nice!(u64, NiceU64, 1);
-inflect_nice!(usize, NiceU64, 1);
-inflect_nice!(NonZeroU8, NiceU8, Self::MIN);
-inflect_nice!(NonZeroU16, NiceU16, Self::MIN);
-inflect_nice!(NonZeroU32, NiceU32, Self::MIN);
-inflect_nice!(NonZeroU64, NiceU64, Self::MIN);
-inflect_nice!(NonZeroUsize, NiceU64, Self::MIN);
-inflect_nice!(i8, NiceU8, 1, unsigned_abs);
-inflect_nice!(i16, NiceU16, 1, unsigned_abs);
-inflect_nice!(i32, NiceU32, 1, unsigned_abs);
-inflect_nice!(i64, NiceU64, 1, unsigned_abs);
-inflect_nice!(isize, NiceU64, 1, unsigned_abs);
+inflect_nice!(3,  u8,           NiceU8,  1);
+inflect_nice!(6,  u16,          NiceU16, 1);
+inflect_nice!(13, u32,          NiceU32, 1);
+inflect_nice!(26, u64,          NiceU64, 1);
+inflect_nice!(26, usize,        NiceU64, 1);
+inflect_nice!(3,  NonZeroU8,    NiceU8,  Self::MIN);
+inflect_nice!(6,  NonZeroU16,   NiceU16, Self::MIN);
+inflect_nice!(13, NonZeroU32,   NiceU32, Self::MIN);
+inflect_nice!(26, NonZeroU64,   NiceU64, Self::MIN);
+inflect_nice!(26, NonZeroUsize, NiceU64, Self::MIN);
+inflect_nice!(3,  i8,           NiceU8,  1,         unsigned_abs);
+inflect_nice!(6,  i16,          NiceU16, 1,         unsigned_abs);
+inflect_nice!(13, i32,          NiceU32, 1,         unsigned_abs);
+inflect_nice!(26, i64,          NiceU64, 1,         unsigned_abs);
+inflect_nice!(26, isize,        NiceU64, 1,         unsigned_abs);
 
 // These aren't nice, but we can still do basic inflection.
 inflect!(u128, 1);
@@ -216,7 +272,7 @@ mod tests {
 		($num:expr, $str:literal) => (
 			t_inflect!($num, $str);
 			assert_eq!(
-				$num.nice_inflect("book", "books"),
+				$num.nice_inflect("book", "books").to_string(),
 				format!(concat!("{} ", $str), $num.to_formatted_string(&Locale::en)),
 				"{}.nice_inflect()", $num
 			);